Refactor tensor display formatting into display module
